perf: extract static Mapbox style metadata to a Mapbox-free module, eliminate unnecessary mapbox-gl bundle on non-map routes
